Rename rss to feed since it supports atom and rss.
Add better support for linting and testing.
