+++ .travis.yml
@@ -10,8 +10,7 @@ python:
 script:
   - pip install pipenv
   - pipenv sync --dev
-  - pipenv run pytest --cov=rss
-  - tests/lint.sh
-  - tests/docstyle.sh
+  - make test
+  - make check-travis
 after_success:
   - coveralls
