+++ app.py
@@ -1,6 +1,6 @@
 """Provides application for development purposes."""
 
-from rss.factory import create_web_app
+from feed.factory import create_web_app
 
 app = create_web_app()
