+++ feed/serializers/__init__.py
@@ -1,0 +1,7 @@
+"""RSS serializers."""
+
+__all__ = ["Atom10", "RSS20"]
+
+
+from feed.serializers.atom_1_0 import Atom10
+from feed.serializers.rss_2_0 import RSS20
