--- rss/serializers/__init__.py
@@ -1,7 +1,0 @@
-"""RSS serializers."""
-
-__all__ = ["Atom10", "RSS20"]
-
-
-from rss.serializers.atom_1_0 import Atom10
-from rss.serializers.rss_2_0 import RSS20
