+++ setup.cfg
@@ -1,5 +1,5 @@
 [flake8]
-ignore = E203,W503
+ignore = E203,E999,W503
 
 [pydocstyle]
 convention = numpy
